+++ performance/components/nested_complex_component.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-class Performance::NameComponent < ViewComponent::Base
+class Performance::NestedComplexComponent < ViewComponent::Base
   def initialize(name:)
-    ("a"..."z").to_a.each do |c|
+    ("a"..."z").to_a.shuffle.each do |c|
       instance_variable_set("@rand_var_#{c}", true)
     end
     @name = name