+++ performance/components/nested_name_component.rb
@@ -2,6 +2,9 @@
 
 class Performance::NestedNameComponent < ViewComponent::Base
   def initialize(name:)
+    ("a"..."z").to_a.each do |c|
+      instance_variable_set("@rand_var_#{c}", true)
+    end
     @name = name
   end
 end