+++ performance/partial_benchmark.rb
@@ -5,6 +5,8 @@
 
 require "benchmark/ips"
 
+Warning[:performance] = true
+
 # Configure Rails Environment
 ENV["RAILS_ENV"] = "production"
 require File.expand_path("../test/sandbox/config/environment.rb", __dir__)
@@ -13,6 +15,8 @@ module Performance
   require_relative "components/name_component"
   require_relative "components/nested_name_component"
   require_relative "components/inline_component"
+  require_relative "components/complex_component"
+  require_relative "components/nested_complex_component"
 end
 
 class BenchmarksController < ActionController::Base
@@ -21,12 +25,15 @@ end
 BenchmarksController.view_paths = [File.expand_path("./views", __dir__)]
 controller_view = BenchmarksController.new.view_context
 
+controller_view.render(Performance::ComplexComponent.new(name: "HI there"))
+
 Benchmark.ips do |x|
   x.time = 10
   x.warmup = 2
 
   x.report("component") { controller_view.render(Performance::NameComponent.new(name: "Fox Mulder")) }
   x.report("inline") { controller_view.render(Performance::InlineComponent.new(name: "Fox Mulder")) }
+  x.report("complex") { controller_view.render(Performance::ComplexComponent.new(name: "Fox Mulder")) }
   x.report("partial") { controller_view.render("partial", name: "Fox Mulder") }
 
   x.compare!